Refactor TypeCastExpressions into separate files
