+++ src/FakeXrmEasy.Core/Query/TypeCastExpressionExtensions/TypeCastExpressionExtensions.EntityReference.cs
@@ -1,0 +1,41 @@
+using System;
+using System.Globalization;
+using System.Linq.Expressions;
+using FakeXrmEasy.Extensions;
+using Microsoft.Xrm.Sdk;
+
+namespace FakeXrmEasy.Query
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    internal static partial class TypeCastExpressionExtensions
+    {
+        internal static Expression GetAppropriateCastExpressionBasedOnEntityReference(Expression input, object value)
+        {
+            Guid guid;
+            if (value is string && !Guid.TryParse((string)value, out guid))
+            {
+                var getNameFromEntityReferenceExpr = Expression.Call(Expression.TypeAs(input, typeof(EntityReference)),
+                    typeof(EntityReference).GetMethod("get_Name"));
+
+                return Expression.Condition(Expression.TypeIs(input, typeof(EntityReference)),
+                    Expression.Convert(getNameFromEntityReferenceExpr, typeof(string)),
+                    Expression.Constant(string.Empty, typeof(string))).ToCaseInsensitiveExpression();
+            }
+
+            var getIdFromEntityReferenceExpr = Expression.Call(Expression.TypeAs(input, typeof(EntityReference)),
+                typeof(EntityReference).GetMethod("get_Id"));
+
+            return Expression.Condition(
+                Expression.TypeIs(input, typeof(EntityReference)),  //If input is an entity reference, compare the Guid against the Id property
+                Expression.Convert(
+                    getIdFromEntityReferenceExpr,
+                    typeof(Guid)),
+                Expression.Condition(Expression.TypeIs(input, typeof(Guid)),  //If any other case, then just compare it as a Guid directly
+                    Expression.Convert(input, typeof(Guid)),
+                    Expression.Constant(Guid.Empty, typeof(Guid))));
+
+        }
+    }
+}
